Implement interface DataRecordsReader
